Tests: fixed hangs with Net::SSLeay with LibreSSL on macOS.
On macOS (as seen on 14.4.1, latest at the moment), Net::SSLeay compiled
with LibreSSL is shipped with the OS (Net::SSLeay 1.88, LibreSSL 3.3.6).
And for some reason mail_ssl_conf_command.t and stream_ssl_conf_command.t
tests hang after the test, waiting indefinitely in reused SSL socket close().

This seems to be an LibreSSL bug in SSL_shutdown() (at least a couple was
fixed in LibreSSL 3.5.0), probably related to session reuse (which do work
in these particular tests due to TLSv1.2 explicitly used for testing).
Still, calling close(SSL_no_shutdown => 1) explicitly is a trivial workaround,
and hence it is implemented.