Tests: optimized processing of large QUIC packets with padding.
Path MTU discovery packets might contain a lot of padding, and creating
a copy of the whole buffer for each PADDING frame, which is just one
byte with type 0, consumes lots of resources.  This was seen to result
in flapping of at least h3_keepalive.t and h3_ssl_early_data.t tests.

Fix is to copy at most 8 bytes for parse_int() calls when parsing
frame types.